Finished adding demo functionality
Though there is no server interaction. Songs can be voted up or down
and voted off if they have a score of -5. One of the songs is preset to
demo this feature. Since that song score is preset, there is a bug in
which it reappears after being voted off. This bug can be avoided
during the demo however.